feat(judgment): organize package structure
