+++ src/main/kotlin/apply/application/LastJudgmentResponse.kt
@@ -1,4 +1,8 @@
-package apply.domain.judgment.tobe
+package apply.application
+
+import apply.domain.judgment.Commit
+import apply.domain.judgment.JudgmentResult
+import apply.domain.judgment.JudgmentStatus
 
 data class LastJudgmentResponse(
     val status: JudgmentStatus,
