+++ src/main/kotlin/apply/domain/judgment/AssignmentArchive.kt
@@ -1,0 +1,7 @@
+package apply.domain.judgment
+
+import java.time.LocalDateTime
+
+interface AssignmentArchive {
+    fun getLastCommit(pullRequestUrl: String, endDateTime: LocalDateTime): Commit
+}
