+++ src/main/kotlin/apply/domain/judgment/Judgment.kt
@@ -1,4 +1,4 @@
-package apply.domain.judgment.tobe
+package apply.domain.judgment
 
 import support.domain.BaseRootEntity
 import java.time.LocalDateTime
