+++ src/main/kotlin/apply/domain/judgment/JudgmentStartedEvent.kt
@@ -1,3 +1,3 @@
-package apply.domain.judgment.tobe
+package apply.domain.judgment
 
 data class JudgmentStartedEvent(val judgmentId: Long, val commit: Commit)
