+++ src/main/kotlin/apply/domain/judgment/JudgmentType.kt
@@ -1,4 +1,4 @@
-package apply.domain.judgment.tobe
+package apply.domain.judgment
 
 enum class JudgmentType {
     EXAMPLE, REAL
