--- src/main/kotlin/apply/domain/judgment/tobe/AssignmentArchive.kt
@@ -1,7 +1,0 @@
-package apply.domain.judgment.tobe
-
-import java.time.LocalDateTime
-
-interface AssignmentArchive {
-    fun getLastCommit(pullRequestUrl: String, limitedDateTime: LocalDateTime): Commit
-}
