+++ src/test/kotlin/apply/AssignmentFixtures.kt
@@ -6,7 +6,7 @@ import apply.application.AssignmentResponse
 import apply.domain.assignment.Assignment
 
 private const val GITHUB_USERNAME: String = "ecsimsw"
-private const val PULL_REQUEST_URL: String = "https://github.com/woowacourse/service-apply/pull/367"
+const val PULL_REQUEST_URL: String = "https://github.com/woowacourse/service-apply/pull/367"
 private const val NOTE = "과제 소감입니다."
 
 fun createAssignment(
