+++ src/test/kotlin/apply/domain/judgment/JudgmentRecordTest.kt
@@ -1,8 +1,9 @@
-package apply.domain.judgment.tobe
+package apply.domain.judgment
 
-import apply.domain.judgment.tobe.JudgmentStatus.FAILED
-import apply.domain.judgment.tobe.JudgmentStatus.STARTED
-import apply.domain.judgment.tobe.JudgmentStatus.SUCCEEDED
+import apply.createJudgmentRecord
+import apply.domain.judgment.JudgmentStatus.FAILED
+import apply.domain.judgment.JudgmentStatus.STARTED
+import apply.domain.judgment.JudgmentStatus.SUCCEEDED
 import io.kotest.assertions.assertSoftly
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
