+++ src/test/kotlin/apply/domain/judgment/JudgmentTest.kt
@@ -1,8 +1,11 @@
-package apply.domain.judgment.tobe
+package apply.domain.judgment
 
-import apply.domain.judgment.tobe.JudgmentStatus.FAILED
-import apply.domain.judgment.tobe.JudgmentStatus.STARTED
-import apply.domain.judgment.tobe.JudgmentStatus.SUCCEEDED
+import apply.createCommit
+import apply.createJudgment
+import apply.createJudgmentRecord
+import apply.domain.judgment.JudgmentStatus.FAILED
+import apply.domain.judgment.JudgmentStatus.STARTED
+import apply.domain.judgment.JudgmentStatus.SUCCEEDED
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
